+++ frontend/src/metabase-types/api/user.ts
@@ -6,6 +6,7 @@ export interface User {
   last_name: string;
   common_name: string;
   email: string;
+  locale: string | null;
   google_auth: boolean;
   is_active: boolean;
   is_qbnewb: boolean;
