+++ frontend/src/metabase/app.js
@@ -12,7 +12,7 @@ import "number-to-locale-string";
 import "metabase/lib/i18n-debug";
 
 // set the locale before loading anything else
-import { loadLocalization } from "metabase/lib/i18n";
+import "metabase/lib/i18n";
 
 // NOTE: why do we need to load this here?
 import "metabase/lib/colors";
@@ -89,16 +89,6 @@ function _init(reducers, getRoutes, callback) {
   initializeEmbedding(store);
 
   store.dispatch(refreshSiteSettings());
-
-  MetabaseSettings.on("user-locale", async locale => {
-    // reload locale definition and site settings with the new locale
-    await Promise.all([
-      loadLocalization(locale),
-      store.dispatch(refreshSiteSettings({ locale })),
-    ]);
-    // force re-render of React application
-    root.forceUpdate();
-  });
 
   PLUGIN_APP_INIT_FUCTIONS.forEach(init => init({ root }));
